branches/main: devicesTutorial example3, use double for task period, not long.
git-svn-id: https://svn.lcsr.jhu.edu/cisst/branches/main@652 0959eb1a-6956-4f60-947b-6092a4e22c97

+++ examples/devicesTutorial/example3/main.cpp
@@ -42,7 +42,7 @@ int main(void)
     cmnClassRegister::SetLoD("mtsTaskManager", CMN_LOG_LOD_VERY_VERBOSE);
     cmnClassRegister::SetLoD("devSensableHD", CMN_LOG_LOD_VERY_VERBOSE);
 
-    const long PeriodDisplay = 1.0; // in milliseconds
+    const double PeriodDisplay = 1.0; // in milliseconds
     mtsTaskManager * taskManager = mtsTaskManager::GetInstance();
     
     displayTask * displayTaskObject = new displayTask("Display", PeriodDisplay * cmn_ms);
